ios 10 fixes

+++ OrigamiEngine/Plugins/HTTPSource.m
@@ -148,7 +148,7 @@ const NSTimeInterval readTimeout = 1.0;
         [data getBytes:buffer length:data.length];
         _bytesRead += data.length;
 
-        result = data.length;
+        result = (int)data.length;
     }
 
     return result;
